wtf is going on

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: wteles-d <wteles-d@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/27 15:18:35 by wteles-d          #+#    #+#             */
-/*   Updated: 2023/05/09 18:08:53 by wteles-d         ###   ########.fr       */
+/*   Updated: 2023/05/16 16:02:21 by wteles-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,9 +25,9 @@
 char	*ft_extract_str(char *str);
 char	*ft_trim_str(char *str);
 char	*get_next_line(int fd);
-int		ft_spec_strcpy(char *dst, const char *src, int i);
+size_t	ft_spec_strcpy(char *dst, const char *src, int i);
 char	*ft_strjoin(char *s1, char *s2);
-char	*ft_strlcpy(char *dst, const char *src, size_t size);
+char	*ft_strlcpy(char *dst, const char *src, int size);
 size_t	ft_strlen(const char *s);
 
 #endif
